Successfully add styled component
