+++ app/electron/preload/webview/elements/insert.ts
@@ -1,18 +1,18 @@
-import { getDeepElement } from './helpers';
+import { CssStyleChange } from '../changes';
+import { getDeepElement, getDomElement } from './helpers';
 
-export function insertElement(x: number, y: number, width: number, height: number) {
+export function insertElement(x: number, y: number, width: number, height: number, tag: string) {
     const el = getDeepElement(x, y) as HTMLElement | undefined;
     if (!el) {
         return;
     }
-    const newDiv = createNewDiv(width, height);
+    const newDiv = document.createElement(tag);
     el.appendChild(newDiv);
-}
 
-function createNewDiv(width: number, height: number) {
-    const newDiv = document.createElement('div');
-    newDiv.style.width = `${width}px`;
-    newDiv.style.height = `${height}px`;
-    newDiv.style.backgroundColor = 'rgba(0, 0, 255, 0.5)';
-    return newDiv;
+    const domEl = getDomElement(newDiv);
+    const change = new CssStyleChange();
+    change.updateStyle(domEl.selector, 'width', `${width}px`);
+    change.updateStyle(domEl.selector, 'height', `${height}px`);
+    change.updateStyle(domEl.selector, 'backgroundColor', 'rgb(120, 113, 108)');
+    return domEl;
 }
